Refactor IMU init and widen timestamps
Split IMU task into init/config functions, add retry/delay/constants,
initialize Madgwick separately and add safety limits for failures.
Use u64 timestamps, refine math (to_degrees, midpoint) and add a
driver type alias. Clean up flash storage serialization (use Self,
add docs, clippy allows) and improve calibration field comments.

+++ src/task/drive/control/rotation.rs
@@ -20,7 +20,7 @@ pub struct RotationState {
     /// Last measured yaw angle (in degrees)
     pub(crate) last_yaw: Option<f32>,
     /// Last update timestamp
-    last_update_ms: u32,
+    last_update_ms: u64,
     /// Direction of rotation
     direction: RotationDirection,
     /// Type of motion during rotation
